AI config engine: rename FlowDiagnostics to FlowResult and move it out of errors.py
FlowDiagnostics was the complete outcome of processing a flow, not an error, so
returning it from Engine.flows out of errors.py was misleading. Rename it to
FlowResult and move it, together with ConfigStatus, into models.py next to
ResolvedFlow; FlowError, ErrorKind, and ConfigError stay in errors.py. Engine.flows
now returns a result type while get_flow() still returns only a valid ResolvedFlow.
